phase2/requirements.txt
# Phase 2 Core Dependencies
torch
transformers
sentence-transformers
scikit-learn
pandas
numpy
fastapi
uvicorn
pillow
python-multipart
joblib
